document_intelligence_wrapper/requirements.txt
azure-ai-documentintelligence==1.0.0b3
pandas==2.2.2
tabulate==0.9.0